feat(web): 封装contentType生成函数
