+++ apps/mis-web/src/utils/file.ts
@@ -13,6 +13,7 @@
 import { stringify } from "csv-stringify";
 import iconv from "iconv-lite";
 import { Encoding } from "src/models/exportFile";
+import { getContentType } from "src/utils/server";
 import { Transform } from "stream";
 
 /**
@@ -40,25 +41,6 @@ export const getCsvStringify = (headerColumns: { [key in string]: string }, colu
   return csvStringify;
 };
 
-// 创建编码转换流到管道
-export const createEncodingTransform = (encoding: Encoding) => {
-  console.log("encoding.toLowerCase()", encoding);
-
-  return new Transform({
-    transform(chunk, _, callback) {
-      let encodedBuffer;
-      if (encoding.toLowerCase() === "gb18030") {
-        encodedBuffer = iconv.encode(chunk.toString(), "gb18030");
-      } else if (encoding.toLowerCase() === "utf-8") {
-        encodedBuffer = chunk.toString("utf-8");
-      } else {
-        encodedBuffer = chunk; // 默认情况下，不改变编码
-      }
-      callback(null, encodedBuffer);
-    },
-  });
-};
-
 type exportKey = "users" | "accounts" | "payRecords" | "chargeRecords" | "operationLogs";;
 
 /**
@@ -81,4 +63,37 @@ export const getCsvObjTransform = (key: exportKey, formatFn: (obj: any) => any) 
       }
     },
   });
+};
+
+// 创建编码转换流到管道
+export const createEncodingTransform = (encoding: Encoding) => {
+  console.log("encoding.toLowerCase()", encoding);
+
+  return new Transform({
+    transform(chunk, _, callback) {
+      let encodedBuffer;
+      if (encoding.toLowerCase() === "gb18030") {
+        encodedBuffer = iconv.encode(chunk.toString(), "gb18030");
+      } else if (encoding.toLowerCase() === "utf-8") {
+        encodedBuffer = chunk.toString("utf-8");
+      } else {
+        encodedBuffer = chunk; // 默认情况下，不改变编码
+      }
+      callback(null, encodedBuffer);
+    },
+  });
+};
+
+// 获取带charset的Content-Type
+export const getContentTypeWithCharset = (filename: string, encoding: Encoding) => {
+  // 获取 MIME 类型
+  const contentType = getContentType(filename, "application/octet-stream");
+  // 添加 charset 信息
+  let contentTypeWithCharset;
+  if (contentType.includes("charset=")) {
+    contentTypeWithCharset = contentType.replace(/charset=[^;]+/, `charset=${encoding}`);
+  } else {
+    contentTypeWithCharset = `${contentType}; charset=${encoding}`;
+  }
+  return contentTypeWithCharset;
 };
